Normalize piece strengths when returning value
This changes divides the strength of each piece so that the
Total_Force_Gene returns 1.0 for the initial board configuration. Before
this change, an increase in the overall value of pieces in the
Piece_Strength_Gene would effectively increase the weight of all other
genes that referenced piece strengths without changing their scalar
values. This change makes the weight of other genes independent of the
Piece_Strength_Gene.