Fix errors in piece_has_moved()
While running a gene pool, some games would end with an exception thrown
from std::map::at(), indicating an attempt to access a key that was not
present. My only guess was that, after a pawn promotion, calling
piece_has_moved() on that piece would generate the error since it was a
new piece and not in the piece_moved std::map. This commit should fix
that problem by making the update to the piece_moved std::map on every
place_piece() call. Since this change, the problem has not shown up.

Since it took several thousand games for this problem to show up, I was
never able to reproduce it. Oh well.