Pawn_Advancement_Gene::score_board() change
Don't take past promotions into consideration. Either it will
overcompensate by making the compensation permenant (go through entire
history) when it is irrelevant in later moves if the promoted piece is
captured, or it will merely push back the drop in score one recursion
depth.

Promoting a pawn will lower the score this gene gives out, but
this should be compensated by an increase in the Total_Force_Gene
output.