Implement multiple gene pools
The problem with the previous setup was that the measure of a Genetic_AI
was if it could be other Genetic_AIs. This works to an extent in the
abscense of an objective standard to measure them by, but it breeds
pathological behavior. To succeed and reproduce, an AI only has to beat
the AIs around it. If those AIs have strange, suboptimal behavior, then
other strange, suboptimal behavior can beat them. Thus, the best AI is
not a good player, but only good at beating bad AIs with bad behavior.
This can be seen in the result graphs on the wiki: closely packed the
entire pool around seemingly randomly.

With multiple gene pools, multiple independent sets of divergently
evolved AIs should present a diverse challenge to any would-be champion
AIs. By periodically transferring clones of the best AIs between pools,
new genes can be introduced and the competition kept honest.

This should improve things as long as the time between interbreeding is
long enough for the evolution of each pool to diverge. There is a
tradeoff between the number of gene pools, the interval between
interbreeding, and the amount of time for a program run.