Fix Board::game_has_ended()
Previous method did not correctly determine if the game on a board had
ended. There are still legal moves when the threefold-repetition rule is
invoked. When a game has really ended, the last entry of the game record
will have a tab character that separates the last move from the result.
This tab character is the flag for the game ending.