Move Board::is_legal() code to Move::is_legal()
The pattern culminating with this commit is that the Board class only
holds the state of the game while the Move classes hold the actions that
can affect the Board state. To use an analogy, the Board is the database
(holding state and validating data) while the Moves are transactions.
The rules of chess are encoded in the Moves, leaving the Board with just
infrastructure.

Also, some #include cleanup.