Show Genetic AI analysis of every legal move
The AI will print the move, predicted future moves, and the score of the board end state (which may be from the opposite perspective) for every legal move in a turn. This should help track down oddities in the decision process.

The large size of this file necessited the clean game record printing method.