Useful utility for timing code
Create an instance of this class anywhere in the code and it will record the time until it reaches end-of-scope. Used to discover that Sphere_of_Influence::score_board() and Board::board_status() were taking far too long to execute.

Speedups allow for up to 200 moves per second analyzed by Genetic AIs.