Simpler King Protection Gene
Instead of counting the unguarded squares from which the other side can attack the king (which takes too long and is quickly deactivated (a guess about why it gets deactivated)), just count the empty squares from which the king can be attacked. This relates to how guarded the king is.