New plot: Piece Strength Gene activity
This new plot shows what percentage of AIs are actually using the Piece
Strength Gene. The gene being in use is defined as when the Piece
Strength Gene is active and at least one other gene that reads it
(e.g., the Total Force Gene) is also active.

The latest gene pool runs don't seem to optimize the piece strength gene
as they did prior to the addition of alpha-beta pruning and the
Board::king_in_check() simplification. In the latest run, only about 30
percent of AIs have an active Piece Strength Gene. The top gene is now
the Sphere of Influence Gene.

My current hypothesis is that the Sphere of Influence Gene still causes
the AIs to favor queens, rooks, bishops, and knights in that order due
to how they can cover the board with their movement. Further evidence:
the queen is still the most popular promotion piece by far. Now that the
Sphere of Influence Gene has a faster calculation due to the faster
Board::king_in_check(), it is favored as more informative (combines
piece strength and position) than the Total Force Gene.