Greatly speed up loading gene pool files
Instead of loading every Genetic_AI in a file and picking out the ones
that were still alive during the last run, find the IDs of the living
AIs and only load those. Obvious, really.

What's less obvious is that this also drastically cuts down on memory
usage, even after the non-living AIs are destroyed after leaving the
load_gene_pool() function scope. The explanation for this is found here:

http://stackoverflow.com/a/34284122/2156537

In short, std::map allocates items on the heap despite it begin on the
stack. This memory cannot always be reclaimed immediately by the OS when
freed, especially after a large allocation (like loading 100k AIs from a
file).