New Look Ahead parameter
Minimum Moves Left specifies the minimum number of moves to leave time for towards the end of the game. This is to prevent an AI from using up the rest of its time on a move that isn't the last.

This gene is starting to seem more like a chromosome.