Precompute all legal moves from all squares for all pieces
In other words, when the Board asks for all legal moves from a piece on
a square, that piece will only return a list of moves that are
potentially legal from that square--"potentially", as in, the move lands
on the board. This cuts down on the number of moves that have to go
through more involved checks for legality (pinned, king in check, etc.).

This results in about a 20% speedup in move searches and fixes #50.