Board::game_record stores Complete_Moves
Fixes #51

Instead of constantly converting moves to their string representation,
store the Complete_Move object and do the parsing only when needed. This
has resulted in a 50% speed-up in move analysis.

Technical note: Inside the Piece class, the possible_moves vector was
made static so that all std::unique_ptr<Move> items are stored there for
the duration of the program run. Before, the possible_moves vector was a
member of each Piece object, and so was destroyed when the piece was
captured on the original board. This caused nullptr dereferencing during
prior attempts at making this change.

Other change: CECP thinking output now prints the speed of move
analysis.