Use correct factor for Piece Strength Gene mutation
In order to use an array for holding piece values, it is necessary that
some array slots are unused since the chars representing the pieces are
not sequential. So, the number of keys needs to be used instead of the
size of the array. This restores this genes' mutation rate to what it
was before the map --> array overhaul.