Use a minimal copy of board for move legality check
In order to make sure that the king is not in check after a move, make a
copy of the board without history information. This is much faster than
making a full copy. Move searching by Genetic_AIs now twice as fast
without losing speed as the repeat_count map gets bigger. The limit on
speed is now (probably) how long it takes to generate the legal moves
list.