More efficient check for legal moves
This commit changes how the board is checked to make sure that the
player who makes a move does not leave their own king in check after the
move. Previously, this was done by copying the board (minimally, without
history) and running the king_is_in_check() method. Since most moves
don't involve the king in anyway, this was not as efficient as it could
have been.

Now, only certain situations require extensive king-check checks: king
already in check, moving piece pinned, en passant (which is always weird
since it captures a piece on a different square than it lands on).
Copying the board is no longer necesssary to check the legality of a
move.

Result: 50% speedup for release build.