Less strict usage of principal variation
Instead of only using the principle variation when the opponent's move goes according to prediction, use it whenever the indicated move is legal. Just because the opponent didn't pick the expected move does not invalidate the rest of the line. All that is needed is to make sure the predicted moves are still legal.