Faster iteration through all attacked sqaures
Using the Threat_Iterator for finding all attacked squares for use in
the King Confinement Gene was 20x slower than on the master branch. This
commit changes the prodedure to cycle through all pieces and get the
attacked squares from the piece moves instead of going through all
squares and searching for any piece that attacks it.

This restores this branch to the same speed as the master branch,
meaning it should be ready to merge soon.

A note about the change to En_Passant: this move is considered to *not*
capture because it does not capture a piece on the square the pawn lands
on. Thus, it is not a threat to the king, which is how special moves are
filtered out. Plus, the change does not affect the result of
Move::is_legal().