New Gene: Priority Threshold
This gene stops a board-scoring gene from running its score_board()
method if the priority of the gene is close enough to zero. This
sacrifices the details of an evaluation for speed. The threshold used
during the Genome::evaluate() method is a random value from zero to the
genetically determined threshold in order to not completely stop the
directed evolution of lower-priority genes.