Replace normal distribution with Laplace distribution
The Laplace distribution (basically, a double-sided exponential distribution) has fatter tails than a normal distribution, so it should result in a greater likelihood of large mutations without turning into chaos.

This should also negate the need for the bit_mutations branch and not require the range of possible values to be pre-determined and hard-coded.