New normalization for King Confinement Gene
Make sure that the the square values in the gene don't suppress the
value of the priority gene. That is, this should fix the problem that,
if the square scores get too high, the priority must come down so the
score assigned by the gene does not overwhelm other genes.

This normalization should keep the score returned by score_board() to
less than 1.