Replace map with array for holding piece hash values
Substantial speed increase from putting piece hash values in a
std::array vs. a std::map. Since the hash values are kept in a static
member variable of Board, I can only conclude that, since the structure
is not copied, that the std::array is cache friendlier than the
std::map.