Better gene pool game management
When the start of a new game in a gene pool is delayed by a limit on the
maximum simultaneous games, start the new game when any currently
running game completes. Previously, the waiting for game completion
would go in order of game start order, missing an opportunity to start a
new game when a later game finished earlier.