New AI: Monte Carlo AI
The idea behind this AI is that, for each legal move in the current
position, random games starting after that move will be played to
completion, keeping track of how many end in wins, draws, and losses.
The move with the highest probability of leading to a win with random
moves is chosen.

Ideas for improvements could be made for choosing random moves:

 - Always pick a move that results in checkmate during the random play
   session.

 - Give moves that alter the material on the board (captures and
   promotions) higher probability of being chosen.