Get rid of floating point comparison warnings
Instead of checking for equality with Math::win_score or between scores
in Game_Tree_Node_Result methods, use std::isinf() for the former
and inequality comparisons for the latter. The entire program now
compiles with no warnings.

An interesting note: this initially led to nonsensical moves during test
games. The problem was that compiling with -Ofast turning on
-ffinite-math-only, which assumes that all numeric values are
finite.This caused problems since the score for checkmate is +infinity.
Luckily, switching from -Ofast to -O3 resulted in no change of speed.
The problem was probably caused by std::isinf() always returning false
due to the assumption of -ffinite-math-only. This caused the Minimax_AI
to continue looking for moves after a checkmate was found, resulting in
illegal moves being returned.