Use C++17 for copy elision for clang warning
Clang warned about having a custom destructor without custom copy or
assignment constructors. As a test, the copy and assignment methods of
Scoped_Stopwatch were explicitly deleted. This lead to an error for the
line in src/Testing.cpp:

    auto perft_watch = Scoped_Stopwatch("");

The error said that this line required the deleted copy constructor.
Changing the standard to C++17 with guaranteed copy elision fixed this.