New feature: legal move count vs. game progress
This function takes a file with PGN listings for multiple games and
outputs a listing of the number of legal moves as each game progresses.
It's a guess of mine that it might be possible to determine the overall
state of the game (opening, midgame, endgame) from the changes in the
number of legal moves.

    1. Both players start with 20 legal moves.
    2. In the opening, pawns move out of the way and other pieces are
    developed, rapidly increasing the number of moves.
    3. The opening phase ends when pieces are traded off and the number
    of legal moves per turn rapidly falls.
    4. The midgame has a mostly constant number of moves (checking
    notwithstanding).
    5. The endgame starts when one player gains an advantage and starts
    capturing more pieces and driving the king into a corner, leading to
    a smaller number of legal moves for one side.