Rremove straight_line_move() from all_empty_between()
The test of all_empty_between() is ill-defined if the two squares are not in line with each other (by row, column, or diagonal). Change the test to an assertion for debugging purposes. The caller of all_empty_between() is now responsible for the squares being queried being in line.