Prevent illegal boards created from FEN strings
All internal Board methods assume that it is impossible for a sequence
of legals moves to result in a king being attacked by more than two
pieces. This commit prevents such board from being created through an
FEN string.

The Square class was giving an operator<() function for use with
std::set in the FEN Board constructor.