Limited recursion depth and improved scoring
In Turing's descirption of his chess computer, scoring is only based on three-ply depth or one extra ply if the move is considerable. I had initially interpretted the description as always recursing to lower depths if a move is considerable.

Also, improve the last_move_captured() function so it doesn't require replaying entire games every time it is called.

Games now complete in a reasonable amount of time. It ends in a three-fold repetition draw after a dozen moves, but it ends.