New elements for Look Ahead Gene
The Look Ahead Gene now has another speculation factor on which to
allocate time: whether or not the previous move in the game captured a
piece (this is in contrast to the previous factor which hinged upon
there being a capturing move available).

Initial results indicate it evolves to the same value as the standard
speculation factor, indicating that the recent past doesn't matter
much.