New scoring algorithm for King Confinement Gene
The final score from this gene is now the weighted sum of squares
blocked by friendly pieces and enemy attacks, all divided by the
number of reqchable free squares. This is a more accurate
representation of the safety of the king since a large number of
free squares is bad when friendly pieces are far away and good when
enemy attacks dominate the board.