Add maximum distance component to King Confinement Gene
Since the King Confinement Gene has the slowest score_board() method,
this change adds a Maximum Distance value that cuts off the calculation for
squares that take more than a set number of king moves to reach. Perhaps
this will give this gene a larger priority compared to the Priority
Threshold Gene.