Fix Game_Tree_Node_Result to prior behavior
The depth of a result only matters in the case of a win or loss.

Even though it's probably rare that two nodes in the game tree end up
with equal scores, it's unclear whether the shallower or deeper result
(game-tree-wise) should prevail in that case. Deeper results have more
of the game behind them and are closer to the end, but are less likely
to actually obtain due to the possibilities of different intervening
moves.