Remove exponents from board-scoring genes
Exponents don't seem to converge to any specific value, except for large
exponents that act equivalently to low priority (x^n is small for large
n and |x| < 1).

A greater number of values to evolve results in slower improvement. It
may be time to start deleting genes that never get significant priority,
both for having a smaller number of gene components to evolve and to
increase the speed of analysis with fewer scoring genes.