Create plots from offspring data
A fit of the child count vs. frequency (i.e., the number of AI that
created N children) is very close to y = (1/2)^N. This indicates that,
for the most part, every AI has a 50/50 chance of winning. The
frequencies for large N tend to be above the fit line. Whether the
interpretation of these points is noise in small numbers or indicators
of very advantageous mutations remains to be seen.