Revert "Remove extra call to Board::piece_on_square() from release build"
This reverts commit 4a7534f0cc7d4f0a5f5a7d19609e5f69e0dbe146.

The reverted commit actually makes the program slower. In fact, here are
a few versions list in order of increasing speed:

    if(attacking_piece)
    {
        if(board.whose_turn() == attacking_piece->color())
        {
            return false;
        }
        ...
    }

=====

    auto attacking_color = board.whose_turn();
    if(attacking_piece)
    {
        if(attacking_color == attacking_piece->color())
        {
            return false;
        }
        ...
    }

=====

    auto moving_piece = board.piece_on_square(starting_file, starting_rank);
    if(attacking_piece)
    {
        if(moving_piece->color() == attacking_piece->color())
        {
            return false;
        }
        ...
    }

The last bit of code is the original that is being reverted to. I admit
I don't understand this behavior. I thought the first would be the
fastest since it avoids any method calls when not needed and the method
merely returns an integer (a Color enum value) instead of calculating an
index and returning an array element.