Prevent reading deleted data
Compiling and running on Windows resulted in read errors when attempting to pull the branches of a lower part of the tree into the top level--i.e., branches = std::move(next_branch->branches). Oddly, the code worked on Linux when compiled with gcc or clang.

The fix here extracts the data to be moved to a higher level (next_branch = std::move(read_branch->branches[index])) before assigning it. That wasy, it is not caught in the recursive destructors before it can be placed in its new location.