Search space limiter for King Confinement Gene
Instead of using  a distance metric like in previous versions, this
limit is based on the number of free (not blocked and not attacked)
squares that the search encounters. This will allow searches along a
confined path (e.g., king is trapped against the back rank by a rook on
the second rank) to proceed farther while not taking too much time in
more open situations (mostly empty board).