Rework opening plots
The python script now filters the data so that only the top 10 of the
opening rate usage data columns get sent to the MATLAB script. Before,
when all of the data was sent, doing plots on the first 2 moves was
impossible since there are about 200,000 possible two move openings and
importing data for 50,000 games overwhelms a desktops memory. Now, the
recent top 10 openings used is determined without having to create a
giant grid of data.