New Gene: Null Gene
This gene acts as a sort of null hypothesis to show what the evolution
of a gene with no affect on Genetic AI behavior looks like. I'm expeting
the priority of this gene to act like a random walk and the resulting
plots to look like a spreading fan of values.