Speedup to Freedom to Move Gene
std::bitset::any() is faster than std::bitset::count(), so check the
former before continuing with the calculation.

This gene's score_board() method is still the slowest, so more
investigation is required. Keeping a count of attacks in the
Board class and updated when the pieces moved has proved to be more
complex, but would be a much better improvement.