Significant speedup of Freedom to Move Gene
The number of attacking moves minus those moves that attack friendly
pieces is now tracked by the Board and updated incrementally after every
move, leading to the Freedom_To_Move_Gene::score_board() to consist of a
single division where prior to this commit it was the slowest scoring
gene. The affect on Board operation speed seems to be negligible.

Since the new algorithm has some complicated logic, the original version
of the scoring method is left as verfication of the workings of the new
one.