Define internal integer types
When using different integer types for the representation of these types, the rules
for promotion can be different between, say, unsigned ints and unsigned shorts.
This is especially true for the Square class which does non-trivial calculations with
the internal integer code for the instance.  The integer promotion rules being
different mean that some tests may fail by changing the type of the square_index.
This commit nails down the types to keep them consistent.