Offspring plot improvements
Instead of fitting a curve to the data, plot the null hypothesis, which
assumes that every Genetic AI has a 50/50 chance of winning. Deviations
from this plot should show that some AIs really are bette than others.
Specifically, an excess of AIs with large numbers of offspring should
mean that substantial chess-playing improvements are being made.

Also, code organization; plus, octave can't put a legend with a scatter
plot, so use the plot() function with extra options.