Improve signal handling during gene pools
On Linux: Simplify pausing/restarting/quitting by using unique_locks
instead of raw mutexes.

On Windows: Signals are not handled the same way as on Linux, so the
only functionality during gene pools is to exit the program either after
the current games are finished (Ctrl-c) or immediately (Ctrl-c twice).