Alan Turing AI now considers depth in scores
If the board being scored is a winner for either side, then the depth of
the look-ahead is considered: shallower wins and deeper losses are
preferred. This prevents loops in the endgame where the side to win
keeps picking moves that are sure to win, but don't bring victory any
closer.

This optimizatoin goes beyond Turing's description, but it is necessary
to get the AI to actually finish games.