Generic function for testing other functions
Unify most tests in Testing.cpp under one generic testing call. This
way, the only action taken on the tests_passed variable is assigning it
false when a test fails. No more confusing &= operations (which are
technically inaccurate as they operate on ints instead of bools).
Results in the deletion of a lot of repetitive boilerplate.