Make sure asserts aren't tripped
During king movements and while the king is temporarily
removed from the board, the assert in the Board::find_king()
method can trip when adjusting the hash after the king's
first move, necessitating the removal of both castling hashes.

Now, to avoid that assert, the king_location array is accessed
directly.