Simplify Board::create_move(string) with <algorithm>s
As it turns out, PGN and coordinate moves are not that far apart
in format. Use a general format of the form PFRFRP, where P is
a piece symbol, F is a file, and R is a rank. The first P is the moving
piece, and the second P is the type of piece a pawn is promoted to.
The first FR is the file and rank where the piece starts, and the second
FR is the file and rank where the piece ends up. Both move formats
drop the redundant information. Coordinate moves use the middle four
characters to specify the move with an optional promotion character,
while PGN uses the first piece symbol and the second FR, optionally
adding the first F and R and the second P as needed to resolve ambiguity.