Replace piece_is_pinned() with in-between check
That is, replacing a searching function with a direct calculation.

Add tests to show why the previous version had if( ! piece_on_square())--
if the king is in check and the moving piece is not capturing the checking
piece, then capturing a piece does not get the king out of check. This is to
prevent allowing a move that captures a piece that blocks check from another
piece that is not now currently checking the king.