Atomically update best genome during gene pool run
In order to have a gene pool running while connected to lichess.org, the
file containing the best genome so far cannot be read by the executable
connected to lichess while it is being written to by the executable
running the gene pool. So, the gene pool will write the file to a random
file name, then use std::filesystem::rename() to atomically replace the
old best genome file.