Derive move that results in new setboard
When playing with a CECP interface on lichess.org, the interface does
not send moves (using "usermove" or otherwise) to update the local
board, but instead sends a "setboard" command to say what the new state
of the board is. This loses information needed for 3-fold repetition
detection. With these changes, if there is a move that causes the
local board to have the same FEN as the newly received board, that move
is applied to the current board in order to keep track of position
repetitions.

Squashed commit of the following:

commit a3434017c82f21aa03646a8fbbd2fa44a03cf3f1
Author: Mark Harrison <mark.z.harrison@gmail.com>
Date:   Thu Jun 6 15:57:50 2019 -0700

    Use simpler setboard parsing

commit cf0b1461cb2addd683dd5408628d64209662cf53
Merge: a135ac4 567cb7e
Author: Mark Harrison <mark.z.harrison@gmail.com>
Date:   Thu Jun 6 15:24:55 2019 -0700

    Merge branch 'master' into derived_move

commit a135ac4f57240e23c28a14cca3fc3701d7f47004
Author: Mark Harrison <mark.z.harrison@gmail.com>
Date:   Wed Jun 5 22:11:48 2019 -0700

    Derive move that results in new setboard

    When playing with a CECP interface on lichess.org, the interface does not use
    "usermove" to update the local board, but instead sends a "setboard"
    command to say what the new state of the board is. This loses information
    needed for 3-fold repetition detection. With this new update, if there is a
    move that causes the local board to have the same FEN as the newly
    received board, that move is applied to the current board.