Print Minimax variation as RAVs in PGN files
Recursive Annotation Variation (RAV) is used for describing variations
from the actual move played. It is used here to print how the Minimax AI
expects the game to go in the near future. The score for the variation is
printed inside a comment at the end of the RAV.

The official description in the PGN standard can be found here:
http://portablegamenotation.com/Specification.html#c8.2.5