Get opponent name from GUI interface
When playing with a GUI interface, the Outside_Communicator learns the
name of the opponent on the other side of the GUI (whether in a
different process or on another computer) and hands it off to the local
player before printing the game record.

Small additional change: In Board::print_game_record(), there no reason
to copy a pointer by const-ref.