Revealed check via en passant problem
Create test to show that a revealed check results in
Board::find_checking_square() failing to find the right square.

Change to Board::find_checking_square() fixes problem by making sure
that if the starting square of the moving piece was not blocking check,
then the search for the checking square starts on the square previously
occupied by the pawn captured en passant.