Make size of priority mutations depend on priority
Get genes with ideally high priorities (especially the Total Force Gene)
to have high priorities faster by making the mutation size equal to the
square root of the priority (minimum 10.0). Perhaps this will allow for
more efficient evolution of lower priority genes since they won't have
to wait ias long for more important genes to be optimized.

The square root was chosen as a compromise between a constant mutation
rate (which leads to linear growth--too slow) and a rate proportional to
the priority (which leads to exponential growth--too fast).