Attempt at not losing AI quality
I've noticed that the quality of the AIs can experience a large
drop after many generations. For example, in the last run,
AIs with IDs in the 100k range could routinely beat AIs with
IDs in the 200k range.

My guess is that, since the AIs are only compared with AIs
in the current generation and AI quality is not necessarily
transitive (A beats B and B beats C does not mean that
A beats C), mutations that improve the probability of victory
in the current generation are not necessarily improvements in
an absolute sense.

This latest change reintroduces random old AIs if they can beat
members of the current generation. The choice is biased towards
later AIs by weighting the probability of choosing an AI to be
proportional to the ID value--picking ID=1000 is ten times as likely
as picking ID=100).