Implement Quiescence Search
The result of implementing quiescence search is that the scores assigned
to the board state as the game progresses are much more stable.
Previously, the score would swing wildly, even in the opening moves, due
to searches stopping just after a major change in the game state
(captures, promotions, etc.) and missing immediate neutralizing or
reversing moves. Now, most immediate consequences of a move are found
before a score is assigned to a series of moves.

Squashed commits (merges from master removed):

Real commits:
commit 61a563a5857c1d29314fefc6ac8e331d93fc9d01
Author: Mark Harrison <mark.z.harrison@gmail.com>
Date:   Fri Aug 30 18:20:57 2019 -0700

    Add quiescent timing to speed test

commit 9f661f305ba15c22f0b57cd4aef9ae24efc87403
Author: Mark Harrison <mark.z.harrison@gmail.com>
Date:   Fri Aug 30 05:07:37 2019 -0700

    Working implementation of Quiescent Search

    https://www.chessprogramming.org/Quiescence_Search

    If, at the leaf of a game tree search, the last move leaves a piece
    vulnerable to capture, then fast-forward the game until all captures on
    the square the piece occupies have occured. This is to prevent the
    evaluation of a board which is, in a sense, unstable due to immanent
    piece capture.

    The captures proceed using the least valuable pieces first and the
    captures are stopped using a simplified minimax procedure to avoid piece
    exchanges that are losing to either player, since a real player would
    stop the captures to avoid too much loss.