Make mutation sizes scale with mutating value
For mutating values with multiple competing values (priorities and piece
strengths here) and where only value ratios matter (queen at 10 and
bishop at 3 will have the same behavior as queen at 100 and bishop
at 30), it can take a long time for values that are properly spaced far
apart to reach that state. To help this along, the size of the mutation
(the width of the Laplace distribution) is scaled by the size of the
square root of the absolute value of the current value. This allows
mutations of large values to still have an effect on game play.

A minimum mutation size is also imposed (10.0 for priorities, 1.0 for
piece strengths) so that values don't get stuck at zero with zero-sized
mutations.