Keep game record separate from Board
To lessen the amount of copying when searching the game tree, this list
of moves made on a Board should not be kept inside the Board structure.
Instead, it should be stored wherever the game is being played.

So far, this only provides a modest speedup in perft tests and none in
GUI-based games.