play_game() takes Clocks by value
Starting with deleting an inaccurate comment in Gene_Pool.cpp, the
"winner" in a drawn game is again picked randomly. There should be
no need to artificially encourage AIs to take more time to choose
moves. This allows play_game() to take the Clock instance by value,
which negates the need to store game clocks in a vector parallel to the
games.