Simplify gene pool writing/parsing
Only write a gene pool generation if new AIs have been created. This
negates the need for checking for consecutive Still ALive lines when
counting the number of rounds so far. Changing a configuration file
before continuing a run (especially the number of pools/AIs) may
cause a miscounting, but this is fine since such changes are usually
followed by restarting a pool anyway. It would only cause off-by-one
errors that are not critical.

Also, trim the whitespace on lines from a gene pool game file in case
of outside edits.