Delete useless smart pointers
Smart pointers do not serve a purpose here. If an exception is thrown,
only the most recently constructed pointer is freed. Every other Move
pointer is leaked. However, the data structure being built is only built
once upon program start, and if it fails, then the whole program fails,
so there's no need to worry about leaked resources.

I could have stored the smart pointers, but that would be a useless
complication since all of this data is supposed to last the life of the
program.