Derive moves from CECP setboard command
If possible, derive the moves made between the current board
state and the state indicated by the CECP setbord command. A
maximum limit of two moves is imposed since it is impossible
in general to tell which of two moves made by the same player
came first. As an example, 1. e4 e5 2. d4 results in the same board
position (with the same FEN) as 1. d4 e5 2. e4.