Handle messages that stop thinking
If an exception is caught (e.g., program receives "quit") while thinking
about a move, then the thinking needs to end immediately so that
the command from the GUI can be acted upon. Similarly, the CECP
command "result" can come at any time to signal the end of a game.
Thinking about a move should stop in this case as well.