Keep game record and board consistent
Found by the assertion at the end of Board::print_game_record().

One of the dangers of keeping the game record separate is that
they have to be updated with new moves in a consistent way. Note
to future self:
Always keep board.submit_move() and game_record.push_back()
(and any other command that modifies both) next to each other.