Fix bug in parsing promotion to bishop in coordinate notation
Since promotion pieces are sent as lowercase letters, the move f7f8b
was initially validated as f7f8b. Since the last character was a lowercase
letter, it was assumed to be a promotion piece and an uppercase
version was appended, resulting in f7f8bB. This is obviously nonsense.

Found by robertomil81mil at lichess.org.