Regularize doxygen comments
Add \brief to the summary of each doxygen block and complete the
comment mark ("\\!") so that spurious comment marks don't appear
in the documented code. The \brief is necessary since otherwise the
brief remark would be grouped with the longer text once the dividing
\\! blank line was included.

Also, inherited members should be inlined so every class has all of its
documentation.