Make sure scoped classes cannot be temporaries
As of now, no C++ compiler allows for [[nodiscard]] to be placed on
constructors. This commit fixes the code so that no temporary variables
can be created of Scoped_Push_Guard or Scoped_Stopwatch by
forcing them to be created through a function.

- The constructors of Fixed_Push_Guard are made private so that only
  the methods of the friend class Fixed_Capacity_Vector can create them.
- Scoped_Stopwatch objects are now created by a static function.

Furthermore, all other construction and assignment methods are deleted
from these classes.