Speed up gene pool loading
Instead of reopening the gene pool file for each AI, open the file
once and keep reading the same stream and only reopen the file
when the end of the file is reached. If an AI is not found when starting
from the beginning the file, then an error is thrown, since the requested
AI does not exist in the file.