Fix gene pool loading speedup
Apparently, Visual Studio is more lenient in accepting a temporary
std::ifstream (from the Genetic_AI file name constructor) as a parameter
taking a std::istream& (non-constant reference). GCC is not accepting of
this, so an rvalue-reference std::istream constructor was added to fix
this. The new read_from() method is used to consolidate duplicate
constructor logic.

Anyway, the result of this work is that loading 48 AIs from a file with
150k now takes 5 seconds instead of 47 seconds.