Faster gene pool loading
Gather all "Still Alive" lines and all live IDs before attempting to
read the actual AI data. Ideally, if all IDs are sorted in the file,
then only two passes through the file are needed: one for the live
IDs and pool numbers, and another for the gene data.

In tests with a file with 200k AIs, loading takes half as long to load
(4 seconds vs. 7 seconds for the previous way).