Silence warning of non-positive log axis limits
In switching between linear and logarithmic axes, the axis limits given
by ylim are not updated (or, perhaps, an implicit lower limit of 0 is
fine, but not when it is set explicitly). This situation does not seem
to be a problem for generated plots, so the warning is silenced as not
helpful.