New Look Ahead aspect: Minimum Fraction of Alpha
This genetic factor controls when a minimax player is allowed to
cut off searching deeper into the game tree. Since alpha represents
the best move found so far in previous branches, it should not be
rejected for a move that has only been analyzed at a much shallower
depth. Therefore, once a value for alpha (as in alpha-beta pruning) has
been established, a game tree search is not allowed to stop until a
depth has been reached that is greater than the given fraction of
the alpha
depth