Delete unnecessary Board checks
The proper functioning of the Board is no longer dependent on the
king having no more than two checking pieces and no more than one
checking knight. While this is a true invariant of a real chess board, it
is not strictly necessary for this program. Deleted code doesn't need
maintenance, so delete it.

For historical reference, this was needed by the Threat_Iterator way
back before there was a Square class.