New method for choosing a best AI to record
The AI with the most wins is considered the best and only AIs that
accumulate more wins will replace it. To make sure that more evolved
AIs get recorded, the number of wins is reduced by a constant factor
(1%) for now. The idea is that wins against older, less-evolved AIs are
worth less than wins against newer, more-evolved AIs.

Hopefully, this will result in a set of AIs with a smaller variability
in quality than the previous method of just taking the oldest living AI
to be the best.