AIs that use more time get to reproduce after draw
It seems to be a common problem that Genetic AIs will settle into a time
usage strategy that leaves a lot of time on the clock (at least for
longer games of several minutes as oposed to several seconds). To
counter this, make the player that used more time in a drawn game the
designated winner for reproductive purposes.