All gene pools should be seeded
When a seed AI is given in the configuration file, all gene pools should
be created from copies of the seed AI. To only have one copy of the seed
in one gene pool is a waste of computing work since the other pools will
be overrun on the first swap. Having only one copy of the seed in each
gene pool results in wasted computation in the games between non-seed
AIs.

Also, adjust the offspring plotting scripts to exclude the seed since it
will have a disproportionate number of offspring since it starts with
multiple copies.