Don't favor variations with late progressive moves
Formerly, the board evaluator would give a slightly higher score to
variations where the pawn/capture move came late in the variation. This
would lead to useless delaying of, for example, an unavoidable pawn
promotion since variations where this move came later resulted in a
higher score due to the (100 - board.moves_since_pawn_or_capture())
factor. Now, variations with a pawn move or a capture are all weighted
the same.