Make sure UCI movestogo affects clock correctly
After the clock is created and after a move is made, the body of the
mehtod Clock::set_next_time_reset() would reference the wrong value in
moves_to_reset_clocks. Punching the clock at the beginning of
UCI_Mediator::setup_turn() fixes this problem by switching the clock to
the right player.