Make reactivation more probable than deactivation
If a gene is deactivated, the time saved from not calling its evaluation
function is a huge block to it becoming reactivated. So, to give
deactivated genes a chance to reactivate, the probability of going from
inactive to active is now ten times larger than going from active to
inactive.