Better name for add_legal_move() parameter
A blockable move is one that cannot bypass another piece on
the board. So, all blockable moves going in the same direction
should be in the same list, ordered by increasing move length.
This allows the loop in Board::recreate_move_caches() to break
on encountering another piece instead of constantly checking
if the move is blocked.