Separate blockable and unblockable attacks
With the introduction of pieces like the Hawk, which can jump
in straight lines, it is no longer possible to always block check
just by putting a piece in between the checking piece and the
king.

The Board class now distinguished between blockable and
unblockable attacks so that extra logic can figure out what
to do.