Specialize Scoped_Stopwatch for timing programs
The Scoped_Stopwatch is now more lightweight due to the
removal of mutexes and any output until calls to
print_all_timings(). A helper method print_guard() guarantees
that no matter how the program exits (short of kill -9), the timings
will be printed just before the program closes.

The parsing of command-line arguments needs to be updated
to handle more than one operation.