Add extra data to plots and analyses
For plots with lots of lines of data, put the final value of the plot
next to the label in the legend to make it easier to see which line is
which.

Also, output a list of current piece values with the queen normalized to
a score of 900. Normally, a pawn would be worth 100 and the units would
be centipawns, but the pawn value that results is usually much smaller,
probably due to the influence of other pawn-centered genes.