Make sure all Piece_Strength_Gene data is recorded
The call to piece_value() with the Piece argument returns zero if the
gene is inactive. For the purposes of writing to a file, the piece
values should be recorded whether or not the gene is active. so, the
piece_value() method taking a Piece_Type should be called instead.