Gated promotions are now like normal moves
Previously, the gated promotion moves would be destructed after every
Board::submit_move() call, leaving dangling pointers in the game record.
Now, all possible promotions are in the normal Move data structure, and
an extra legality check is added to Pawn_Promotion to make sure the
promoted piece is allowed on the board.