Stop minimax search on repeat board position
During a search, if the current variation includes a loop, then the
variation is assigned a draw score and not pursued further. The
search should not continue down that line even though threefold
repetition has not been reached because that branch will be
identical to the branch starting from the previous duplicate board
aside from the ply count.

If the best line goes through repeated board positions, then the best
line will go through that board position again, leading to a threefold
repetition. Thus, a draw score is assigned.