Change piece strengths writing for better plots
In a standard chess piece scoring system, a pawn is worth 1 point, a
queen 9, a rook 5, and bishops and knights 3. The total score of a full
set of pieces at the beginning of the game is thus 39 (the king is
ignored because it cannot be captured). When writing to a file,
normalize the piece strengths so that the total value of a full set of
pieces (ignoring the king) is 39. This solves the problem of a possible
division by zero when normalizing to a single piece value. Also, it is
now much easier to compare the genetically determined values of chess
pieces versus the standard values given above.

In the gene_plots.m script, round piece strengths to 2 decimal places
and simplify the use of these values.