Raise the scaling factor in Freedom To Move Gene
Since there are many board positions in which the number of legal moves
is greater than 20, the score returned by this gene was often greater
than 1. With a higher average score than other genes, the evolved
priority was pushed towards zero. One could tell that this gene was
still valuable because any offspring that ended up with a negative
priority for this gene was quickly eliminated.

With this commit, the score returned by the gene is now less than 1 for
nearly all positions. To get a higher number of legal moves, you need
crazy positions like this (which has 218 legals moves):

    R6R/3Q4/1Q4Q1/4Q3/2Q4Q/Q4Q2/pp1Q4/kBNN1KB1 w - - 0 1