Simplify Look Ahead Gene speculation
The introduction of the branching factor/minimum depth calculation seems
to have made the distinction between default and material change
speculation constants irrelevant. When the constants have different
values, it is often the material change constant that is lower, unlike
before. This commit is a try at seeing what happens when only the number
of moves controls how far the AI looks ahead.