AI Mutation is externally controlled
Instead of letting the mutation rate be subject to mutation,
it is now determined in the gene pool configuration file. As
the gene pool runs, it switches between a high and low rate
according to the user's preference. The expected way for
this to be used is for long periods of slow mutation to be
punctuated with short periods of fast mutation. The slow
periods optimize while the fast periods allow bigger changes
to pass over local chess fitness minima to another possible
maxima.

Closes #76